refactor: Update import paths for CropVarietyEntity and related queries in crop modules
- Changed import paths for CropVarietyEntity to reflect its new location in the crops-variety module.
- Updated tests in various command and query handlers to utilize the correct import paths for CropVarietyEntity and related queries.
